pop copy paste buttons and a few fixes

+++ modules/parsing/map.cpp
@@ -1483,7 +1483,7 @@ border_cutoff = 1100.0
                     file << "\t\tuses_potentials = yes\n";
                 }
 
-                file << "\t\tcolor = {" << commodity.r << commodity.g << commodity.b << "}\n";
+                file << "\t\tcolor = {" << (int)commodity.r << (int)commodity.g << (int)commodity.b << "}\n";
 
                 file << "\t}\n";
             }
